add rails-i18n gem; prepare some more translations; show view for melange; show errors on forms;
