+++ app/controllers/melanges_controller.rb
@@ -2,14 +2,13 @@ class MelangesController < ApplicationController
   load_and_authorize_resource
 
   def show
-
   end
 
   def new
   end
 
   def create
-    response = CreateMelange.call(@melange, current_user)
+    response = Melanges::CreateMelange.call(@melange, current_user)
     if response.success?
       redirect_to melange_path(response.response_object), notice: I18n.t("controllers.melanges.create.ok")
     else
@@ -21,11 +20,15 @@ class MelangesController < ApplicationController
   end
 
   def update
-
+    response = Melanges::UpdateMelange.call(@melange, melange_params)
+    if response.success?
+      redirect_to melange_path(response.response_object), notice: I18n.t("controllers.melanges.create.ok")
+    else
+      render :edit, status: 422
+    end
   end
 
   def destroy
-
   end
 
   def index
