+++ app/javascript/packs/flatpickr.js
