+++ app/views/melanges/edit.html.erb
@@ -3,11 +3,14 @@
 </div>
 
 <%= simple_form_for @melange do |form| %>
+  <%= form.error_notification %>
+  <%= form.error_notification message: form.object.errors[:base].to_sentence if form.object.errors[:base].present? %>
+
   <div class="row">
     <div class="col-12">
       <div class="card shadow mb-4">
         <div class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
-          <h6 class="m-0 font-weight-bold text-primary"><%= t("activeview.melanges.#{action_name}") %></h6>
+          <h6 class="m-0 font-weight-bold"><%= t("activeview.melanges.#{action_name}") %></h6>
         </div>
         <div class="card-body">
           <div class="row">
